Fix bug, add filttration and trimming of bars
