+++ app/datasets/training/extract_annotations_from_mscore_svg.py
@@ -2,6 +2,7 @@ import json
 from svgelements import *
 from statistics import median
 from ...Annotation import Annotation
+from .process_bar_lines import process_bar_lines
 
 def extract_annotations_from_mscore_svg(
     from_path_svg: str, 
@@ -134,32 +135,3 @@ def process_staff_lines(
             staff = []
     
     return staves
-
-def process_bar_lines(
-    bar_lines: list,
-    annotation_bboxes: list,
-    staves: list[Annotation]
-):
-    # Sort the bar lines by their x and then y coordinate (final order from left to right and from top to bottom)
-    bar_lines_sorted = sorted(bar_lines, key=lambda x: x.bbox()[0])
-    bar_lines_sorted.sort(key=lambda x: x.bbox()[1])
-    
-    TWO_BARLINE_DIFF = 20   # There can be a measure, that ends with two bar lines, but the space between them is not another measure.
-    NEW_STAFF_DIFF = 1.1 * staves[0].height     # The minimal space between the staves must be slightly greater than the height of the staff
-
-    # Create a bounding box for each bar
-    staff_index = 0
-    for i in range(len(bar_lines_sorted)):
-        annotation_class = Annotation.CLASSES[2]
-        x, y, _, _ = bar_lines_sorted[i].bbox()
-        x2, y2, _, _ = bar_lines_sorted[i+1].bbox() if i+1 < len(bar_lines_sorted) else (0, 0, 0, 0)
-
-        if (y2 - y) > NEW_STAFF_DIFF: 
-            staff_index += 1
-            continue   # This is not a measure, but a new line = new staff. (Or the end of the score.)
-        if (x2 - x) < TWO_BARLINE_DIFF: continue    # This is not a measure, but a double bar line.
-
-        width = x2 - x
-        height = staves[staff_index].height
-        annotation_bbox = Annotation(annotation_class, int(x), int(y), int(width), int(height))
-        annotation_bboxes.append(annotation_bbox.to_json())
